push, each 갱신 연산자
document 속 배열에 push 할 수 있다.
each로 반복을 돌 수 있다.
List를 넘겨도 되긴 하는데 차이점은?

+++ src/main/kotlin/io/taesu/springdatamongodb/app/converters/AppConverters.kt
@@ -28,11 +28,11 @@ class AppConverters(converters: List<Converter<*, *>>) {
 }
 
 @MongodbReadingConverter
-class YearReadConverter: Converter<Int, Year> {
-    override fun convert(source: Int): Year = Year.of(source)
+class YearReadConverter: Converter<Long, Year> {
+    override fun convert(source: Long): Year = Year.of(source.toInt())
 }
 
 @MongodbWritingConverter
-class YearWriteConverter: Converter<Year, Int> {
-    override fun convert(source: Year): Int = source.value
+class YearWriteConverter: Converter<Year, Long> {
+    override fun convert(source: Year): Long = source.value.toLong()
 }
